Cleanup and deduplication on bands
It was noticed that a lot of transmitters were using the same
combinations of bands. Therefore instead of having entries for each band
used by each transmitter, an intermediate table "bandgroups" was
created, i.e. each transmitter is associated to one bandgroup (but the
number of bandgroups is actually very low since many transmitters are
very similar). Considering the number of entries in "bandgroups" (less
than 2400 entries, i.e. only 2x the amound of id_bands), it was
considered realistic to lower the complexity and move the frequencies
directly within "bandgroups" and avoid table id_bands completely.